feat(#81): add javadocs and one more puzzle
