--- src/test/java/it/XMLOutput.java
@@ -1,56 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016-2023 Objectionary.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package it;
-
-import com.jcabi.xml.XML;
-import com.jcabi.xml.XMLDocument;
-import java.io.ByteArrayOutputStream;
-import java.io.OutputStream;
-import org.cactoos.Output;
-
-/**
- * XML output.
- */
-public class XMLOutput implements Output {
-
-    final ByteArrayOutputStream baos;
-
-    public XMLOutput() {
-        this(new ByteArrayOutputStream());
-    }
-
-    private XMLOutput(final ByteArrayOutputStream baos) {
-        this.baos = baos;
-    }
-
-
-    @Override
-    public OutputStream stream() {
-        return this.baos;
-    }
-
-    public XML xml() {
-        return new XMLDocument(this.baos.toByteArray());
-    }
-}
